// src/common/enums/response-messages.enum.ts
export enum ResponseMessagesEnums {
  SUCCESS = 'SUCCESS',
  NOT_AUTHENTICATED = 'NOT_AUTHENTICATED',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  CANNOT_CAST_THIS_ACTION_TO_YOU = 'CANNOT_CAST_THIS_ACTION_TO_YOU',
  CANT_CREATE_MORE_BUSINESS = 'CANT_CREATE_MORE_BUSINESS',
  VERIFY_YOUR_BUSINESS_TO_CREATE_NEW_ONE = 'VERIFY_YOUR_BUSINESS_TO_CREATE_NEW_ONE',
  INVALID_DATA_PROVIDED = 'INVALID_DATA_PROVIDED',
  BUSINESS_NOT_FOUND = 'BUSINESS_NOT_FOUND',
  BUSINESS_NOT_FOUND_ON_USER = 'BUSINESS_NOT_FOUND_ON_USER',
  YOU_CANT_DO_THIS_ACTION_TO_YOURSELF = 'YOU_CANT_DO_THIS_ACTION_TO_YOURSELF',
  YOU_ARE_NOT_THE_OWNER_OF_THIS_COMPANY = 'YOU_ARE_NOT_THE_OWNER_OF_THIS_COMPANY',
  YOU_ARE_NOT_A_MEMBER_OF_THIS_COMPANY = 'YOU_ARE_NOT_A_MEMBER_OF_THIS_COMPANY',
  THIS_USER_NOT_BELONG_ON_YOUR_COMPANY = 'THIS_USER_NOT_BELONG_ON_YOUR_COMPANY',
  ONE_TIME_ONLY_ACTION = 'ONE_TIME_ONLY_ACTION',
  NOT_FOUND = 'NOT_FOUND',
  ALREADY_EXIST = 'ALREADY_EXIST',
  THIS_USER_ALREADY_EXIST_IN_YOUR_COMPANY = 'THIS_USER_ALREADY_EXIST_IN_YOUR_COMPANY',
  THIS_INVITE_DONT_BELONG_TO_YOU = 'THIS_INVITE_DONT_BELONG_TO_YOU',
  THIS_INVITE_IS_NOT_VALID = 'THIS_INVITE_IS_NOT_VALID',
  NO_PERMISSION_FOR_THIS_ACTION = 'NO_PERMISSION_FOR_THIS_ACTION',
  SUBSCRIPTION_PLAN_ID_NOT_FOUND = 'SUBSCRIPTION_PLAN_ID_NOT_FOUND',
  USERS_NOT_IN_RELATION = 'USERS_NOT_IN_RELATION',
  AN_ACTIVE_PLAN_IS_NEEDED_FOR_THIS_ACTION = 'AN_ACTIVE_PLAN_IS_NEEDED_FOR_THIS_ACTION',
}